+++ rpc/src/main/java/com/walmartlabs/concord/rpc/ProcessHeartbeatService.java
@@ -1,0 +1,8 @@
+package com.walmartlabs.concord.rpc;
+
+import com.walmartlabs.concord.sdk.ClientException;
+
+public interface ProcessHeartbeatService {
+
+    void ping(String instanceId) throws ClientException;
+}