+++ rpc/src/main/java/com/walmartlabs/concord/rpc/RunnerApiClient.java
@@ -12,6 +12,7 @@ import java.util.concurrent.TimeUnit;
 public class RunnerApiClient {
 
     private final ManagedChannel channel;
+    private final ProcessHeartbeatService heartbeatService;
     private final KvService kvService;
     private final SecretStoreService secretStoreService;
     private final EventService eventService;
@@ -22,6 +23,7 @@ public class RunnerApiClient {
                 .usePlaintext(true)
                 .build();
 
+        this.heartbeatService = new ProcessHeartbeatServiceImpl(channel);
         this.kvService = new KvServiceImpl(channel);
         this.secretStoreService = new SecretStoreServiceImpl(channel);
         this.eventService = new EventServiceImpl(channel);
@@ -50,5 +52,9 @@ public class RunnerApiClient {
 
     public SlackService getSlackService() {
         return slackService;
+    }
+
+    public ProcessHeartbeatService getHeartbeatService() {
+        return heartbeatService;
     }
 }