+++ server/api/src/main/java/com/walmartlabs/concord/server/api/process/ProcessStatus.java
@@ -10,5 +10,6 @@ public enum ProcessStatus {
     RESUMING,
     FINISHED,
     FAILED,
-    CANCELLED
+    CANCELLED,
+    STALLED
 }