+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentCommand.java
@@ -54,6 +54,7 @@ public class AgentCommand implements Serializable {
 
     public enum Status {
         CREATED,
-        SENT
+        SENT,
+        FAILED
     }
 }