+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentManager.java
@@ -2,8 +2,11 @@ package com.walmartlabs.concord.server.agent;
 
 import com.walmartlabs.concord.rpc.CancelJobCommand;
 import com.walmartlabs.concord.server.api.process.ProcessEntry;
+import com.walmartlabs.concord.server.api.process.ProcessStatus;
 import com.walmartlabs.concord.server.process.queue.ProcessQueueDao;
 import com.walmartlabs.concord.server.rpc.CommandQueueImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -11,6 +14,8 @@ import java.util.UUID;
 
 @Named
 public class AgentManager {
+
+    private static final Logger log = LoggerFactory.getLogger(AgentManager.class);
 
     private final ProcessQueueDao queueDao;
     private final CommandQueueImpl commandQueue;
@@ -24,13 +29,13 @@ public class AgentManager {
     public void killProcess(UUID instanceId) {
         ProcessEntry e = queueDao.get(instanceId);
         if (e == null) {
-            // TODO throw an exception?
-            return;
+            throw new IllegalArgumentException("Process not found: " + instanceId);
         }
 
         String agentId = e.getLastAgentId();
         if (agentId == null) {
-            // TODO throw an exception?
+            log.warn("killProcess ['{}'] -> trying to kill a process w/o an agent", instanceId);
+            queueDao.update(instanceId, ProcessStatus.CANCELLED);
             return;
         }